Handled table index getters and setters.
We don't (yet) maintain variables which have a constant value and what that
value is. This means, that when getting an index using []-syntax, the
resulting type is unkown (i.e. top).

When setting something to some type T with []-syntax, I add T as a possible
type for each already present table member, and make sure the table ends with
top.